+++ src/index.js
@@ -1,14 +1,12 @@
 /* eslint-disable max-len */
 import * as hasMethods from './has/index';
 import * as isMethods from './is/index';
-import * as matchMethods from './match/index';
 import * as meetsMethods from './meets/index';
 import * as multiMethods from './multi/index';
 import * as noMethods from './no/index';
 // Export our methods so they can be tree shaken when someone brings in simply_valid
 export * from './has/index';
 export * from './is/index';
-export * from './match/index';
 export * from './meets/index';
 export * from './multi/index';
 export * from './no/index';
@@ -26,20 +24,15 @@ const extend = (...args) => args.reduce((acc, x) => {
 // Our collection of validation methods extend them so we get their methods and thats it
 
 export const simplyValid = (methodArr, options) => {
-	const methods = extend(hasMethods, isMethods, matchMethods, meetsMethods, noMethods, multiMethods);
+	const methods = extend(hasMethods, isMethods, meetsMethods, noMethods, multiMethods);
 	// Set our default options that can be overwritten if needed.
 	const defaults = {
 		max: Infinity,
 		min: -Infinity,
-		maxLength: 20,
-		minLength: 1,
-		basePattern: '',
-		antiPattern: '',
 		vinPattern: /^[a-hj-npr-z0-9]{9}[a-hj-npr-tv-y1-9]{1}[a-hj-npr-z0-9]{7}$/i,
 		emailPattern: /^[\w\u00c0-\u017f][\w.-_\u00c0-\u017f]*[\w\u00c0-\u017f]+[@][\w\u00c0-\u017f][\w.-_\u00c0-\u017f]*[\w\u00c0-\u017f]+\.[a-z]{2,4}$/i,
 		// Minimum eight characters, at least one uppercase letter, one lowercase letter, one number and one special character
-		passwordPattern: /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[$@$!%*?&])[A-Za-z\d$@$!%*?&]{8,}$/,
-		equalTo: ''
+		passwordPattern: /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[$@$!%*?&])[A-Za-z\d$@$!%*?&]{8,}$/
 	};
 
 	// Collect and set a good portion of options to keep a global set