+++ src/match/index.js
@@ -1,3 +1,0 @@
-export const matchesPattern = (val, {basePattern}) => basePattern.test(val);
-
-export const doesNotMatch = (val, {antiPattern}) => !antiPattern.test(val);